18-projects/2-cloudtaskhub/requirements.txt
# fastapi requirements
fastapi
uvicorn[standard]
httpx

# kafka
aiokafka

# opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-aiokafka

# prometheus
prometheus-fastapi-instrumentator

